[buildsys: hotfix update]
Update Makefile to be able to compile from scratch without a dependency
to a bc-1.06 clone.
Also replace the install target with a simple symlink.

The entire build system is setup very wrong. All the generated files
should be removed from the repo, such that the ./configure && make cycle
works again.